+++ lib/legion/extensions/pushbullet.rb
@@ -1,6 +1,11 @@
 # frozen_string_literal: true
 
 require 'legion/extensions/pushbullet/version'
+require 'legion/extensions/pushbullet/helpers/client'
+require 'legion/extensions/pushbullet/runners/push'
+require 'legion/extensions/pushbullet/runners/device'
+require 'legion/extensions/pushbullet/runners/channel'
+require 'legion/extensions/pushbullet/client'
 
 module Legion
   module Extensions